Process filenames in alphabetical order.
Since most log files are named in a fashion that causes them to sort by date
automatically, this allows the program to not have to do any complex reordering
of its own (which wouldn't be as memory-efficient).

A solution for the more general case of unordered files may be desirable at some
point.

Signed-off-by: Jason Lynch <jason@calindora.com>