By default, trust the timestamp given by a session stop event in LogReader.
The reason for previously ignoring these is that there is a good chance they are
incorrect, as they are generally based on loss of connection to the server,
which may occur some time after the client stops receiving messages. (Depending
on the server and client configurations, this can even be several hours long.)

However, logs generated in chatstats format are presumably generated by clients
that are aware of these issues and mitigate them accordingly. Once readers of
other log formats are written, they may need to ignore session stop events,
depending on their reliability in that particular log format. (mIRC's, for
example, are not known for their reliability.)

Ultimately, this only matters for determining log coverage (that is, how much of
the given timespan is actually covered by a log).

Signed-off-by: Jason Lynch <jason@calindora.com>