Add a new operation to find words that occur with a particular frequency.
This operation is actually quite esoteric, and it may not have a whole lot of
general use. Its purpose is finding words that occur with a particular frequency
for use in a secret word game. It lists words that occur near the specified
frequency on average (specified in average number of seconds between occurances)
and excludes words that ever have a gap between occurances greater than eight
times the specified average.

It's possible that with additional parameterization, this may be a useful
operation, but for the moment, it's being included merely so I don't have to
maintain two separate source trees.

Signed-off-by: Jason Lynch <jason@calindora.com>