Add the ability to manually link nicknames.
This commit allows the user to specify a user information file via the
--users-file command-line option, which (at the moment) contains information
regarding the nicknames each user has used.

The nicknames are subsequently linked by the stats generation code into a single
user for stats purposes, allowing a users' multiple nicknames to be combined
into a single entry in the generated stats.

Signed-off-by: Jason Lynch <jason@calindora.com>