Ignore short gaps in coverage for maintaining the userhost cache.
The userhost cache is built by examining any join event and tracking nick
changes. Every time the log session ends and a new one starts, we are forced to
clear the entire cache, as there is the possibility of a nick's userhost
changing in that time span. However, this results in a marked increase of
undefined userhosts when trying to link nicknames.

This patch helps mitigate this somewhat by only clearing the cache if the gap in
log coverage was longer than five minutes. The duration is completely arbitrary
and should probably be converted to an option at some point.

Signed-off-by: Jason Lynch <jason@calindora.com>