Process specific user specifictions before general ones.
Previously, the order of processing wildcard-laden user specifications from the
users file was in an undefined order. This meant a more general specification,
such as '*!*@*.com' could trump a more specific one such as '*!*@*.aol.com'.
Now, the users are processed in order from most specific to most general,
allowing overlap to be handled naturally.

The order for time-restricted nicknames, however, is based solely on the order
in the users file. This may be inconvienent, but shouldn't matter in practice,
since time-ranges should be used with full hosts rarely, if ever.

Signed-off-by: Jason Lynch <jason@calindora.com>